+++ all-is-cubes/src/block/eval/tests.rs
@@ -189,7 +189,7 @@ fn from_voxels_zero_bounds() {
 fn overall_color_ignores_interior() {
     let resolution = R8;
     let outer_bounds = GridAab::for_block(resolution);
-    let inner_bounds = outer_bounds.expand(FaceMap::repeat(-1));
+    let inner_bounds = outer_bounds.shrink(FaceMap::repeat(1)).unwrap();
     let outer_color = Rgba::new(1.0, 0.0, 0.0, 1.0);
     let inner_color = Rgba::new(0.0, 1.0, 0.0, 1.0);
     let voxels = Evoxels::from_many(